backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiogram
python-dotenv
